app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
motor = "^3.5.1"
fastapi = "^0.112.2"
unicorn = {extras = ["standard"], version = "^2.0.1.post1"}
cmake = "^3.30.2"
pydantic-settings = "^2.4.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
colorama = "^0.4.6"
termcolor = "^2.4.0"
boto3 = "^1.35.12"
uvicorn = "^0.30.6"
mysql-connector-python = "^9.0.0"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
bcrypt = "^4.2.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
